# Implement json.Marshaler in OrderedMap