Resolves https://github.com/wk8/go-ordered-map/issues/13