@fenollp 2022-12-04
@Albert221 Thanks for this! I need it as well. Do you mind adding `Unmarshal` as well? Pretty please