@Albert221 2022-12-05
Benchmark before any changes:

```
goos: darwin
goarch: arm64
pkg: github.com/wk8/go-ordered-map/v2
BenchmarkMarshalJSON-8            440414              2720 ns/op            1473 B/op         58 allocs/op
PASS
ok      github.com/wk8/go-ordered-map/v2        2.371s
```

Benchmark after changing concats to `strings.Builder`:

```
goos: darwin
goarch: arm64
pkg: github.com/wk8/go-ordered-map/v2
BenchmarkMarshalJSON-8            683971              1765 ns/op             472 B/op         15 allocs/op
PASS
ok      github.com/wk8/go-ordered-map/v2        2.501s
```

Benchmark after changing concats to `bytes.Buffer`:

```
goos: darwin
goarch: arm64
pkg: github.com/wk8/go-ordered-map/v2
BenchmarkMarshalJSON-8            720588              1643 ns/op             352 B/op         11 allocs/op
PASS
ok      github.com/wk8/go-ordered-map/v2        2.496s
```

Benchmark after using `strconv.FormatInt` instead of `fmt.Sprintf`:

```
goos: darwin
goarch: arm64
pkg: github.com/wk8/go-ordered-map/v2
BenchmarkMarshalJSON-8           7273260              1650 ns/op             352 B/op         11 allocs/op
PASS
ok      github.com/wk8/go-ordered-map/v2        13.985s
```