@wk8 2022-12-07
@Albert221 thanks for the patch, added a commit on top and merged in https://github.com/wk8/go-ordered-map/pull/16 :)